backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# LangChain & LangGraph
langchain==0.1.6
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph==0.0.20
langsmith==0.0.87

# Database & Async Drivers
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1
sqlalchemy==2.0.25

# Redis
redis==5.0.1
aioredis==2.0.1

# Snowflake
snowflake-connector-python==3.6.0
# Note: snowflake-sqlalchemy removed - not used in codebase and conflicts with SQLAlchemy 2.x
# If needed in future, use version compatible with SQLAlchemy 2.x

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0

# Observability & Monitoring
structlog==24.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0

# Testing
pytest>=8.2.0  # Updated: pytest-asyncio requires >=8.2
pytest-asyncio>=0.24.0  # Updated: supports pytest 8.x (0.23.4 only supports <8)
pytest-cov==4.1.0
# httpx is already listed in Utilities section above
locust==2.20.1

# Development
black==24.1.1
ruff==0.1.15
mypy==1.8.0
pre-commit==3.6.0
